# docs/swagger.yaml
basePath: /api
definitions:
  domain.Car:
    properties:
      id:
        type: string
      marca:
        type: string
      model:
        type: string
      price:
        type: integer
    type: object
  domain.User:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: jimmyvb16@gmail.com
    name: Jimmy Valdez
  description: This is a small CRUD in Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Car API
  version: "1.0"
paths:
  /v1/cars/all:
    get:
      consumes:
      - application/json
      description: get all car.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Car'
            type: array
      summary: get all car
      tags:
      - Car
  /v1/cars/create:
    post:
      consumes:
      - application/json
      description: create a new car.
      parameters:
      - description: Create Car Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Car'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create a new car
      tags:
      - Car
  /v1/cars/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete one car.
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: delete one car
      tags:
      - Car
  /v1/cars/find/{id}:
    get:
      consumes:
      - application/json
      description: get one car.
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Car'
      summary: get one car
      tags:
      - Car
  /v1/cars/update/{id}:
    put:
      consumes:
      - application/json
      description: update one car.
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Car Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Car'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: update one car
      tags:
      - Car
  /v1/login:
    post:
      consumes:
      - application/json
      description: Login for get access token.
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Login for get access token
      tags:
      - Token
  /v1/register:
    post:
      consumes:
      - application/json
      description: Create a new User.
      parameters:
      - description: Create User Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: create a new user
      tags:
      - Token
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"