# internal/core/domain/user.go
package domain

type User struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

type UserResponse struct {
	ID       string `json:"id"`
	Username string `json:"username"`
	JWT      string `json:"token"`
}